storage: consolidate how we handle kafka errors
